Provide service account information when doing an upgrade
