+++ src/ServiceControl.Config/UI/InstanceAdd/ServiceControlAddViewModel.cs
@@ -160,16 +160,7 @@
 
         public void UpdateFromInstance(ServiceControlInstance instance)
         {
-            var userAccount = UserAccount.ParseAccountName(instance.ServiceAccount);
-            UseSystemAccount = userAccount.IsLocalSystem();
-            UseServiceAccount = userAccount.IsLocalService();
-            UseProvidedAccount = !(UseServiceAccount || UseSystemAccount);
-
-            if (UseProvidedAccount)
-            {
-                ServiceAccount = instance.ServiceAccount;
-            }
-
+            SetupServiceAccount(instance);
             InstanceName = instance.Name;
             HostName = instance.HostName;
             PortNumber = instance.Port.ToString();
@@ -258,16 +249,7 @@
 
         public void UpdateFromInstance(ServiceControlAuditInstance instance)
         {
-            var userAccount = UserAccount.ParseAccountName(instance.ServiceAccount);
-            UseSystemAccount = userAccount.IsLocalSystem();
-            UseServiceAccount = userAccount.IsLocalService();
-            UseProvidedAccount = !(UseServiceAccount || UseSystemAccount);
-
-            if (UseProvidedAccount)
-            {
-                ServiceAccount = instance.ServiceAccount;
-            }
-
+            SetupServiceAccount(instance);
             InstanceName = instance.Name;
             Description = instance.Description;
             HostName = instance.HostName;
